refactor(storefront): extract GameApi and StoreApi into components/store/types
GameApi lived inside GameListItem.tsx while six other modules imported it
for reasons unrelated to that component. Move it to a dedicated types
module alongside a new StoreApi, and re-export from GameListItem so any
straggler import keeps working.
