+++ pages/search/index.tsx
@@ -7,7 +7,8 @@ import { Search, SlidersHorizontal, ArrowUpDown, Tag } from "lucide-react";
 
 import { CATEGORIES } from "lib/games";
 import { StoreLayout } from "components/store/StoreLayout";
-import { GameListItem, type GameApi } from "components/store/GameListItem";
+import { GameListItem } from "components/store/GameListItem";
+import { type GameApi } from "components/store/types";
 
 export default function SearchPage() {
   const router = useRouter();
